make the bottomsheet half and full expanded
